wip bundle remap test
